requirements.txt
streamlit
face_recognition
pillow
opencv-python
